+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Unreleased
 
+-
+-
+
+
+# 0.7.1-SNAPSHOT
+
+- addded - `put!`
+- changed/fixed - resume takes a map `(resume {:url url})`
+- changed - default-storage-url "/tmp/db.sqlite"
 
 # 0.7.0 (2025-08-28)